implement BrowserProfiler.sample method using popen.  dumb, straightforward, but fast enough to make a point.

+++ src/native_service/sample.h
@@ -4,6 +4,6 @@
  */
 #ifndef __SAMPLE_H__
 
-bool get_sample(int pid, int &residentMem, int &cpuUsage);
+bool get_sample(int pid, long long &residentMem, float &cpuUsage);
 
 #endif
